+++ CHANGES.md
@@ -1,6 +1,16 @@
 # Release History and TODO
 
 
+## 0.9.0 / 2016-July-??
+
+* Drop support for JDK 1.5 (now Java 6 or higher is required)
+* Drop support for Clojure `1.2` through `1.4` (for code cleanup)
+* Drop support for DSL API for JDBC params (use Cumulus instead)
+* Drop support for `:val-query` option (use `:test-query` instead)
+* Use Apache DBCP2 instead of DBCP1.x
+  * API options for DBCP2
+
+
 ## 0.8.2 / 2016-July-14
 
 * Refactor DSL for constructing basic JDBC params